[FEAT] user image 저장 로직 수정

+++ src/main/java/com/fourback/bemajor/domain/image/service/ImageService.java
@@ -1,16 +1,12 @@
 package com.fourback.bemajor.domain.image.service;
 
-import com.fourback.bemajor.domain.image.entity.Image;
-import com.fourback.bemajor.domain.image.entity.UserImage;
+import com.fourback.bemajor.domain.image.entity.ImageEntity;
 import com.fourback.bemajor.domain.image.repository.ImageRepository;
-import com.fourback.bemajor.domain.user.entity.UserEntity;
 import com.fourback.bemajor.domain.user.repository.UserRepository;
-import com.fourback.bemajor.global.exception.ExceptionEnum;
-import com.fourback.bemajor.global.exception.kind.NotFoundElementException;
+import com.fourback.bemajor.global.common.service.ImageFileService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.core.io.Resource;
 import org.springframework.core.io.UrlResource;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -20,7 +16,6 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 
 @RequiredArgsConstructor
 @Service
@@ -30,30 +25,17 @@ public class ImageService {
     private final UserRepository userRepository;
     private final ImageFileService imageFileService;
 
-
-    public Image findByFileName(String fileName) {
-        Optional<Image> oi = imageRepository.findByFileName(fileName);
-        if (oi.isEmpty()) {
-            throw new NotFoundElementException(1, "That is not in DB", HttpStatus.NOT_FOUND);
-        }
-        return oi.get();
-    }
-
-    @Transactional
     public Resource get(String fileName) throws IOException {
         Path filePath = Paths.get(UPLOAD_DIR).resolve(fileName).normalize();
         return new UrlResource(filePath.toUri());
     }
 
     @Transactional
-    public void saveImage(Image image, MultipartFile file) throws IOException {
+    public void save(ImageEntity image, MultipartFile file) throws IOException {
         if (file != null) {
-            String originalFilename = file.getOriginalFilename();
-            String extension = originalFilename.substring(originalFilename.lastIndexOf('.'));
-            String uniqueFilename = UUID.randomUUID().toString() + extension;
-            Path filePath = Paths.get(UPLOAD_DIR, uniqueFilename);
-            imageFileService.saveImageFile(filePath, file);
-            image.setFilePath(filePath.toString());
+            String uniqueFilename = imageFileService.saveImageFile(file);
+            String filePath = UPLOAD_DIR+uniqueFilename;
+            image.setFilePath(filePath);
             image.setFileName(uniqueFilename);
         }
         imageRepository.save(image);
@@ -62,21 +44,15 @@ public class ImageService {
     @Transactional
     public void delete(List<String> fileNames) throws IOException {
         for (String fileName : fileNames) {
-            Image image = findByFileName(fileName);
+            Optional<ImageEntity> oi = imageRepository.findByFileName(fileName);
+            if(oi.isEmpty()) continue;
+            ImageEntity image = oi.get();
             imageFileService.deleteImageFile(image.getFilePath());
             imageRepository.delete(image);
         }
     }
 
-    @Transactional
-    public String save(Long userId, MultipartFile file) throws IOException {
-        UserEntity userEntity = userRepository.findById(userId).orElseThrow(()
-                -> new NotFoundElementException(ExceptionEnum.NOTFOUNDELEMENT.ordinal(),
-                "This is not in DB", HttpStatus.LOCKED));
-        UserImage image = new UserImage();
-        userEntity.setUserImage(image);
-        image.setUser(userEntity);
-        this.saveImage(image, file);
-        return image.getFileName();
+    public List<ImageEntity> getImagesById(Long postId){
+        return imageRepository.findByPostId(postId);
     }
 }
